+++ TractorMod/release-notes.md
@@ -4,6 +4,7 @@
 * Fixed slingshot enabled by default. (This doesn't affect existing `config.json` files.)
 * Fixed scythe not harvesting bushes in garden pots.
 * Fixed tea bushes affected by `HarvestForage` option instead of `HarvestCrops`.
+* Fixed compatibility with Yet Another Harvest With Scythe Mod (thanks to bcmpinc!).
 * Improved translations. Thanks to therealmate (added Hungarian)!
 
 ## 4.9.2